requirements.txt
flask
request
jsonify
torch
torchvision
torchaudio
transformers
